refactor: reorganize type imports to streamline type management and reduce issues
