+++ src/types/messages/index.ts
@@ -7,11 +7,12 @@
 import { SENSITIVITY_VALUES } from 'constants/index';
 import type { MailsEditorV2 } from 'types/editor/index.d';
 import { Participant } from 'types/participant/index.d';
+import { PrefsType } from 'types/settings';
 import { SaveDraftResponse, MessageSignature } from 'types/soap/index.d';
 
-type MailAuthenticationHeader = { value: string; pass: boolean };
+export type MailAuthenticationHeader = { value: string; pass: boolean };
 
-type MailAuthenticationHeaders = {
+export type MailAuthenticationHeaders = {
 	dkim?: MailAuthenticationHeader;
 	spf?: MailAuthenticationHeader;
 	dmarc?: MailAuthenticationHeader;
