+++ src/types/search/index.ts
@@ -9,7 +9,7 @@ import type { QueryChip } from '@zextras/carbonio-search-ui';
 import { SortBy } from '@zextras/carbonio-ui-commons';
 
 import { NormalizedConversation } from 'types/conversations';
-import { IncompleteMessage, MailMessage } from 'types/messages/index.d';
+import { IncompleteMessage, MailMessage } from 'types/messages';
 import { ErrorType, SearchRequestStatus } from 'types/state/index.d';
 import { AdvancedFilterModalFormValues, Query } from 'views/search/types/types';
