+++ src/views/app/detail-panel/preview/parts/info-block/mail-authentication-header-icon.tsx
@@ -8,7 +8,7 @@ import React from 'react';
 import { Icon, Row, Tooltip } from '@zextras/carbonio-design-system';
 import { useTranslation } from 'react-i18next';
 
-import { MailAuthenticationHeaders } from 'types/index.d';
+import { MailAuthenticationHeaders } from 'types/messages';
 import {
 	getAuthenticationHeadersIconColor,
 	getMailAuthenticationHeaderLabel
