+++ src/views/app/detail-panel/preview/parts/info-details-modal/subsections/mail-authentication-headers-subsection.tsx
@@ -8,7 +8,7 @@ import React from 'react';
 import { Container, Divider, Icon, Padding, Row, Text } from '@zextras/carbonio-design-system';
 import { useTranslation } from 'react-i18next';
 
-import { MailAuthenticationHeaders } from 'types/index.d';
+import { MailAuthenticationHeaders } from 'types/messages';
 import { MailAuthenticationHeaderRow } from 'views/app/detail-panel/preview/parts/info-details-modal/subsections/mail-authentication-header-row';
 import { getAuthenticationHeadersIconColor } from 'views/app/detail-panel/preview/parts/utils';
